+++ src/util.rs
@@ -1,6 +1,8 @@
 use bitcoin;
 use bitcoin::blockdata::script;
 use bitcoin::Script;
+
+use {ScriptContext, ToPublicKey};
 pub(crate) fn varint_len(n: usize) -> usize {
     bitcoin::VarInt(n as u64).len()
 }
@@ -20,4 +22,27 @@ pub(crate) fn witness_to_scriptsig(witness: &[Vec<u8>]) -> Script {
         }
     }
     b.into_script()
+}
+
+// trait for pushing key that depend on context
+pub(crate) trait MsKeyBuilder {
+    /// Serialize the key as bytes based on script context. Used when encoding miniscript into bitcoin script
+    fn push_ms_key<Pk, Ctx>(self, key: &Pk) -> Self
+    where
+        Pk: ToPublicKey,
+        Ctx: ScriptContext;
+}
+
+impl MsKeyBuilder for script::Builder {
+    fn push_ms_key<Pk, Ctx>(self, key: &Pk) -> Self
+    where
+        Pk: ToPublicKey,
+        Ctx: ScriptContext,
+    {
+        if Ctx::is_tap() {
+            self.push_slice(&key.to_x_only_pubkey().serialize())
+        } else {
+            self.push_key(&key.to_public_key())
+        }
+    }
 }